+++ blueprints/starter/navigation/impl/src/commonMain/kotlin/software/ralf/app/platform/template/navigation/NavigationDetailRenderer.kt
@@ -1,0 +1,48 @@
+package software.ralf.app.platform.template.navigation
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import software.ralf.app.platform.inject.ContributesRenderer
+import software.ralf.app.platform.renderer.ComposeRenderer
+import software.ralf.app.platform.template.navigation.NavigationDetailPresenter.Model
+
+@ContributesRenderer
+class NavigationDetailRenderer : ComposeRenderer<Model>() {
+  @Composable
+  override fun Compose(model: Model, modifier: Modifier) {
+    Column(
+      modifier = modifier.fillMaxSize().background(MaterialTheme.colorScheme.background),
+      verticalArrangement = Arrangement.Center,
+      horizontalAlignment = Alignment.CenterHorizontally,
+    ) {
+      Text(
+        text = "Hello, welcome to vRallev/app-platform Template App",
+        style = MaterialTheme.typography.headlineMedium,
+        textAlign = TextAlign.Center,
+        modifier = Modifier.padding(16.dp),
+      )
+      Text(
+        text = "Every 3 seconds a new exampleValue is generated: ${model.exampleValue}",
+        style = MaterialTheme.typography.bodyLarge,
+        textAlign = TextAlign.Center,
+        modifier = Modifier.padding(16.dp),
+      )
+      Text(
+        text = "Total number of exampleValues shown: ${model.exampleCount}",
+        style = MaterialTheme.typography.bodyLarge,
+        textAlign = TextAlign.Center,
+        modifier = Modifier.padding(16.dp),
+      )
+    }
+  }
+}